chore: address comments on pr
