--- src/utils/__mocks__/fileSystem.ts
@@ -1,8 +1,0 @@
-import { vi } from 'vitest';
-
-export const readFile = vi.fn();
-export const writeFile = vi.fn();
-export const copyFile = vi.fn();
-export const exists = vi.fn();
-export const rmdir = vi.fn();
-export const mkdir = vi.fn();
